refactor(types): simplify type exports (#10243)
