+++ .eslintignore
@@ -1,4 +1,3 @@
 dist
 playground-temp
 temp
-packages/vite/client/types.d.ts
