+++ .gitignore
@@ -5,7 +5,6 @@
 *.local
 *.log
 /.vscode/
-/packages/vite/client/types.d.ts
 /packages/vite/LICENSE
 dist
 dist-ssr
