--- packages/vite/src/client-types.d.ts
@@ -1,24 +1,0 @@
-export type {
-  CustomEventMap,
-  InferCustomEventPayload,
-  InvalidatePayload
-} from './types/customEvent'
-export type {
-  HMRPayload,
-  ConnectedPayload,
-  UpdatePayload,
-  Update,
-  PrunePayload,
-  FullReloadPayload,
-  CustomPayload,
-  ErrorPayload
-} from './types/hmrPayload'
-export type { ModuleNamespace, ViteHotContext } from './types/hot'
-export type {
-  ImportGlobOptions,
-  GeneralImportGlobOptions,
-  KnownAsTypeMap,
-  ImportGlobFunction,
-  ImportGlobEagerFunction
-} from './types/importGlob'
-export type { ImportMetaEnv, ImportMeta } from './types/importMeta'
