--- packages/vite/src/dep-types/package.json
@@ -1,3 +1,0 @@
-{
-  "//": "this file is just here to make pnpm happy with --frozen-lockfile"
-}
