+++ packages/vite/src/types/alias.d.ts
