+++ packages/vite/src/types/chokidar.d.ts
