+++ packages/vite/src/types/commonjs.d.ts
