+++ packages/vite/src/types/dynamicImportVars.d.ts
