+++ packages/vite/src/types/http-proxy.d.ts
