+++ packages/vite/src/types/shims.d.ts
