+++ packages/vite/src/types/terser.d.ts
