+++ playground/hmr/event.d.ts
@@ -1,16 +1,9 @@
-import 'vite'
-import 'vite/client/types'
+import 'vite/types/customEvent'
 
-interface MyCustomEventMap {
-  'custom:foo': { msg: string }
-  'custom:remote-add': { a: number; b: number }
-  'custom:remote-add-result': { result: string }
-}
-
-declare module 'vite' {
-  interface CustomEventMap extends MyCustomEventMap {}
-}
-
-declare module 'vite/client/types' {
-  interface CustomEventMap extends MyCustomEventMap {}
+declare module 'vite/types/customEvent' {
+  interface CustomEventMap {
+    'custom:foo': { msg: string }
+    'custom:remote-add': { a: number; b: number }
+    'custom:remote-add-result': { result: string }
+  }
 }
